non_rel_db/serializers.py
from rest_framework import serializers

from .models import Employee, Student


class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = '__all__'


class StudentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Student
        fields = '__all__'
        extra_kwargs = {'id': {'read_only': True}}
